django_api/requirements.txt
Django==5.0.2
django-debug-toolbar==3.4.0
django-split-settings==1.2.0
python-dotenv==1.0.0
psycopg2==2.9.9
djangorestframework==3.14.0
uwsgi